refactor: hook the new Search up to the application layer
Hook up the new Search with KitsuSearch to the application layer. Currently the ViewModel will not
use the response and will need to be fixed in a later PR
