+++ app/src/main/java/com/chesire/nekome/injection/ServerModule.kt
@@ -11,8 +11,6 @@ import com.chesire.nekome.kitsu.api.auth.KitsuAuthService
 import com.chesire.nekome.kitsu.api.library.KitsuLibraryService
 import com.chesire.nekome.kitsu.api.library.LibrarySeriesModelAdapter
 import com.chesire.nekome.kitsu.api.library.ParsedRetrieveResponseAdapter
-import com.chesire.nekome.kitsu.api.search.KitsuSearchService
-import com.chesire.nekome.kitsu.api.search.SearchSeriesModelAdapter
 import com.chesire.nekome.kitsu.interceptors.AuthInjectionInterceptor
 import com.chesire.nekome.kitsu.interceptors.AuthRefreshInterceptor
 import com.squareup.moshi.Moshi
@@ -96,27 +94,5 @@ object ServerModule {
             .addConverterFactory(MoshiConverterFactory.create(moshi))
             .build()
             .create(KitsuLibraryService::class.java)
-    }
-
-    /**
-     * Builds and provides the instance of [KitsuSearchService].
-     */
-    @Provides
-    @Reusable
-    fun providesSearchService(httpClient: OkHttpClient): KitsuSearchService {
-        val moshi = Moshi.Builder()
-            .add(ImageModelAdapter())
-            .add(SeriesStatusAdapter())
-            .add(SeriesTypeAdapter())
-            .add(SearchSeriesModelAdapter())
-            .add(SubtypeAdapter())
-            .build()
-
-        return Retrofit.Builder()
-            .baseUrl(KITSU_URL)
-            .client(httpClient)
-            .addConverterFactory(MoshiConverterFactory.create(moshi))
-            .build()
-            .create(KitsuSearchService::class.java)
     }
 }
