+++ features/search/src/main/java/com/chesire/nekome/app/search/SearchViewModel.kt
@@ -5,14 +5,14 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.chesire.nekome.core.AuthCaster
+import com.chesire.nekome.core.Resource
 import com.chesire.nekome.core.extensions.postError
 import com.chesire.nekome.core.extensions.postLoading
 import com.chesire.nekome.core.extensions.postSuccess
 import com.chesire.nekome.core.flags.AsyncState
 import com.chesire.nekome.core.flags.SeriesType
 import com.chesire.nekome.core.models.SeriesModel
-import com.chesire.nekome.core.Resource
-import com.chesire.nekome.server.api.SearchApi
+import com.chesire.nekome.search.api.SearchApi
 import com.hadilq.liveevent.LiveEvent
 import kotlinx.coroutines.launch
 
@@ -50,7 +50,8 @@ class SearchViewModel @ViewModelInject constructor(
                 else -> error("Unexpected series type provided")
             }
 
-            parseSearchResponse(response)
+            // TODO:
+            //parseSearchResponse(response)
         }
     }
